lxp: Stage 2 memory-safety fixes (32-bit overflow / OOB / fd hygiene) + tests
Surgical fixes for a bug class mostly invisible to the 64-bit host suite because it
wraps a 32-bit size_t (the target is 32-bit); several are reachable from a hostile or
buggy guest:

- poll: cap nfds at LXP_MAX_FDS before nfds*sizeof(pollfd) (2a).
- writev: cap iovcnt at IOV_MAX before iovcnt*sizeof(iovec) (2b).
- pwrite (tmpfs): reject off+len wrap before the reserve + memcpy (2c).
- arena: fail arena_align / lxp_arena_alloc when the size rounding wraps (2d); guard
  the mmap2 text-share pgoff*4096(+len) overflow.
- symlink: validate the target pointer (stored verbatim + strlen'd) via user_strnlen (2e).
- kill/tkill/tgkill: range-check the signal (0..NSIG-1) before it indexes
  sig_handler[] / pending_sig (2f, coordinator).
- dup/dup2/dup3: clear FD_CLOEXEC on the new fd; dup3(fd, fd) -> EINVAL (2g).
- proc_open: release the /proc content slot when fd_alloc fails (2h). (pty is
  scan-reclaimed — open_master reuses no-live-ends slots — so its fd_alloc-fail path
  self-heals, no hook needed.)
- fd_alloc: zero .nonblock (was stale from the slot's prior occupant) (2i).
- tmpfs: reject need > pool before the 256-round wraps (2j).
- evdev: widen the per-open ring tail to uint32_t to match g_in_head (a uint16_t
  desynced after 65536 events) (2k).

Adds tests/suites/test_overflow.c: 5 host regression tests (poll/writev bounds, dup
cloexec, dup3(fd,fd), symlink target fault) that assert the guards and, under the
Stage-1 ASan/UBSan build, scan OOB pre-fix.

Verified: host cmocka 59/59 (plain + ASan/UBSan clean); gate-rot 7/7; QEMU M1/M2/M3;
check-decoupled OK.

+++ include/lxp/lxp_dev.h
@@ -102,7 +102,7 @@ struct lxp_dev_open {
 			uint16_t addr;
 		} i2c; /**< I2C_SLAVE address (P2 i2c). */
 		struct {
-			uint16_t tail;
+			uint32_t tail; /* matches the uint32_t g_in_head; a uint16_t wrapped after 65536 events */
 			uint8_t overrun;
 		} input; /**< evdev ring cursor (P4 input). */
 	} u;
